Remove pxf-service dependency in pxf-hive
Moved `isUserImpersonationEnabled` to the Utilities class in pxf-api.
